AI_service/requirements.txt
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.2
crewai==0.1.0
pandas==2.1.4
numpy==1.24.3
yfinance==0.2.28
ta==0.10.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
asyncio==3.4.3
aiohttp==3.9.1
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.8 